Port part of v1 runner logic into v2 runner
